requirements.txt
streamlit
tensorflow
numpy
Pillow
h5py